基于复杂网络的社交关系可视化分析系统-1/requirements.txt
Flask==2.0.1
numpy>=1.21.0
pandas>=1.3.0
networkx>=2.8.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
pyyaml==6.0
Werkzeug==2.0.3
torch>=1.12.1
torch-geometric>=2.3.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
networkx>=2.8.0
flask>=2.0.0
flask-cors>=3.0.0
torchvision==0.13.1
torchaudio==0.12.1
pymysql>=1.0.2